requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
gunicorn==21.2.0
jwt==1.3.1
packaging==24.0
psycopg2==2.9.7
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
